Better handling of encodings on ruby 1.9.
This makes parsed ApeItems always use a US-ASCII encoding for keys
and a UTF-8 encoding for values.  In addition, when you update
a tag, it will transcode the key and values to US-ASCII and UTF-8,
respectively.  This should be complete encoding support on ruby 1.9.

Add a couple new tests and a few new test data files related
to this support.