Make ApeTag_get_fields and ApeTag_raw take pointer to size
Whenever allocating memory automatically, the size allocated
should be set so the user doesn't have to call a separate
function to determine it.

For ApeTag_raw, the man page recommends using a macro that is
now private to get the size, so in that case there was no
longer a good way to get the size.

Also, for both of these functions, change the handling to
set the pointed to values to NULL and 0 first, and only for
sucessful completion set the actual pointer and size values.

While here, add a check for there being fewer items in the
database that num_fields, so that a partially filed array
won't be returned in that case, since that could easily lead
to crashes later.