ApeTag_get_items() returns a NULL-terminated array
This ensures we always return a non-NULL pointer in the
case of an itemless tag, reserving NULL for error return
values.

A NULL item_count is also now allowed as the user no longer
needs to care about keeping track of the array size/position
while iterating.  This can make it slightly easier to iterate
through items without needing to keep track of the total count
in a separate variable.