Always return -1 on error, add errcode to struct ApeTag
Instead of using -1, -2, and -3 to express the type of error
that occurred, always return -1 on errors, and store the
error type in the ApeTag struct, which the user can retreive
via ApeTag_error_code.  Add multiple different types of errors,
so it's easy for a user to use a switch statement to detect
and act appropriately on different error types.